follow flex parameter specification change
Gbp-Pq: bison-follow-parameter-handling-changes.patch.

+++ lib/scanner-smi.h
@@ -21,7 +21,7 @@
 #define YY_NO_UNPUT
 
 extern char *yytext;
-extern int yyleng;
+extern size_t yyleng;
 
 extern int smiEnterLexRecursion(FILE *file);
 extern void smiLeaveLexRecursion();
